Time: 161 ms (7.19%) | Memory: 126.8 MB (5.06%) - LeetSync

+++ leetcode/334-increasing-triplet-subsequence/increasing-triplet-subsequence.java
@@ -3,17 +3,17 @@ class Solution {
         if (nums.length < 3) return false ;
 
         int n = nums.length;
-        long [] nge = new long [n];
-        long [] lse = new long [n];
-        Stack <Long> st = new Stack<>();
+        int [] nge = new int [n];
+        int [] lse = new int [n];
+        Stack <Integer> st = new Stack<>();
 
         for (int i = 0; i < n ; i++ ){
             while (!st.isEmpty() && nums[i] <= st.peek()) st.pop();
 
-            if (st.isEmpty()) lse[i] = Long.MIN_VALUE;
+            if (st.isEmpty()) lse[i] = Integer.MIN_VALUE;
             else lse[i] = st.peek();
             
-            st.push((long) nums[i]);
+            st.push(nums[i]);
         }
 
         while (!st.isEmpty()) st.pop();
@@ -21,15 +21,15 @@ class Solution {
         for (int i = n-1; i >= 0; i-- ){
             while (!st.isEmpty() && nums[i] >= st.peek()) st.pop();
 
-            if (st.isEmpty()) nge[i] = Long.MIN_VALUE;
+            if (st.isEmpty()) nge[i] = Integer.MIN_VALUE;
             else nge[i] = st.peek();
             
-            st.push((long) nums[i]);
+            st.push(nums[i]);
         }
     
 
         for (int i = 0; i < n; i++){
-            if (lse[i] != Long.MIN_VALUE && nge[i] != Long.MIN_VALUE)return true;
+            if (lse[i] != Integer.MIN_VALUE && nge[i] != Integer.MIN_VALUE)return true;
         }
